[FEAT] 트랜잭션 강화
