+++ src/main/java/com/project/catxi/common/auth/service/TokenService.java
@@ -47,15 +47,6 @@ public class TokenService {
     private final MemberRepository memberRepository;
     private final RefreshTokenRepository refreshTokenRepository;
     private final TokenBlacklistRepository tokenBlacklistRepository;
-    private final KakaoAccessTokenRepository kakaoAccessTokenRepository;
-    private final KakaoFeignClient kakaoFeignClient;
-    private final MatchHistoryRepository matchHistoryRepository;
-    private final ChatMessageRepository chatMessageRepository;
-    private final ChatRoomRepository chatRoomRepository;
-    private final ChatParticipantRepository chatParticipantRepository;
-    private final KickedParticipantRepository kickedParticipantRepository;
-    private final ReportRepository reportRepository;
-    private final DeleteLogRepository deleteLogRepository;
 
 
     //reissue
@@ -148,121 +139,6 @@ public class TokenService {
 
     public boolean isNNDuplicate(String nickname) {
         return memberRepository.existsByNickname(nickname);
-    }
-
-    public void resignation(String accessToken) {
-        try {
-            // 1. JWT 검증 & 사용자 식별
-            if (!jwtUtil.validateToken(accessToken)) {
-                throw new CatxiException(MemberErrorCode.INVALID_TOKEN);
-            }
-
-            Claims claims = jwtUtil.parseJwt(accessToken);
-            String email = jwtUtil.getEmail(claims);
-            Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new CatxiException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-            // 2. 카카오 연결 해제 시도 (실패해도 회원 탈퇴는 진행)
-            boolean kakaoUnlinked = unlinkKakaoAccount(email);
-
-            if (!kakaoUnlinked) {
-                log.warn("⚠️ 카카오 연결 해제 실패, 회원 탈퇴만 진행: {}", email);
-            }
-
-            // 3. DB 정리
-            dropMemberData(member, accessToken, email);
-            
-        } catch (CatxiException e) {
-            log.error("❌ 회원 탈퇴 실패: {}", e.getMessage());
-            throw e;
-        } catch (Exception e) {
-            log.error("❌ 회원 탈퇴 처리 오류 발생: {}", e.getMessage());
-            throw new CatxiException(MemberErrorCode.WITHDRAWAL_FAILED);
-        }
-    }
-
-    //카카오 연결 해제
-    private boolean unlinkKakaoAccount(String email) {
-        try {
-            String kakaoAccessToken = kakaoAccessTokenRepository.findByEmail(email)
-                .orElse(null);
-                
-            if (kakaoAccessToken == null) {
-                log.warn("⚠️ 카카오 액세스 토큰을 찾을 수 없음: {}", email);
-                return false;
-            }
-
-            kakaoFeignClient.unlinkUser("Bearer " + kakaoAccessToken);
-            kakaoAccessTokenRepository.delete(email);
-            log.info("✅ 카카오 연결 해제 및 토큰 삭제 완료: {}", email);
-            return true;
-
-        } catch (FeignException e) {
-            log.error("❌ 카카오 연결 해제 실패: {}", e.contentUTF8());
-            return false;
-        } catch (Exception e) {
-            log.error("❌ 카카오 연결 해제 처리 중 오류: {}", e.getMessage());
-            return false;
-        }
-    }
-
-    //DB 정리
-    @Transactional
-    protected void dropMemberData(Member member, String accessToken, String email) {
-        try {
-            //1. AccessToken 블랙리스트 등록
-            addAccessTokenToBlacklist(accessToken);
-            
-            //2. RefreshToken 삭제
-            refreshTokenRepository.delete(email);
-
-            //3.멤버 삭제 로그화
-            deleteLogRepository.save(DeleteLog.builder()
-                .deletedEmail(email)
-                .memberId(member.getId())
-                .createdAt(LocalDateTime.now())
-                .build());
-
-            //3. 연관 엔티티들 삭제
-            
-            // 3-1. 신고 기록 삭제 (reporter, reportedMember)
-            reportRepository.deleteAllByReporter(member);
-            reportRepository.deleteAllByReportedMember(member);
-            
-            // 3-2. 매치 히스토리 삭제
-            matchHistoryRepository.deleteAllByUser(member);
-            
-            // 3-3. 채팅 참가자 기록 삭제
-            chatParticipantRepository.deleteAllByMember(member);
-            
-            // 3-4. 강퇴된 참가자 기록 삭제
-            kickedParticipantRepository.deleteAllByMember(member);
-            
-            // 3-5. 채팅 메시지 삭제
-            chatMessageRepository.deleteAllByMember(member);
-            
-            // 3-6. 호스트로 생성한 채팅룸 삭제
-            chatRoomRepository.deleteAllByHost(member);
-
-            // 4. 최종 회원 삭제 (Hard Delete)
-            memberRepository.delete(member);
-            log.info("✅ 회원 DB삭제 완료: {}", email);
-
-        } catch (Exception e) {
-            log.error("❌ 회원 데이터 정리 실패: {}", e.getMessage());
-            throw new CatxiException(MemberErrorCode.WITHDRAWAL_FAILED);
-        }
-    }
-
-    //남은 액세스 토큰 블랙리스트 등록(연결 차단)
-    private void addAccessTokenToBlacklist(String accessToken) {
-        Claims claims = jwtUtil.parseJwt(accessToken);
-        Date expiration = claims.getExpiration();
-        long remainTime = expiration.getTime() - System.currentTimeMillis();
-        
-        if (remainTime > 0) {
-            tokenBlacklistRepository.addTokenToBlacklist(accessToken, Duration.ofMillis(remainTime));
-        }
     }
 
     //리프레시토큰 검증
